move out featureExtractors and subjectChangers in own api files
moreover:
- move functions which are not deprecated but in deprecated....kt files
  to the corresponding non-deprecated file (announced BC break)
