+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/ArraySubjectChangerSamples.kt
@@ -274,7 +274,7 @@ class ArraySubjectChangerSamples {
                     toContain(false)
 
                     // still evaluated even though above `toContain` already fails
-                    //                 use `.asList().` if you want a fail fast behaviour
+                    // use `.asList().` if you want a fail fast behaviour
                     toContain.inAnyOrder.atLeast(3).value(true)
                 }
         }
