+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/charSequenceToContainCheckers.kt
@@ -1,8 +1,3 @@
-//TODO rename file to charSequenceToContain... in 0.18.0
-//TODO remove JvmMultifileClass with 1.0.0
-@file:JvmMultifileClass
-@file:JvmName("CharSequenceContainsCheckersKt")
-
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.api.infix.en_GB.creating.charsequence.contains.impl.StaticNames
@@ -10,8 +5,6 @@ import ch.tutteli.atrium.logic._logic
 import ch.tutteli.atrium.logic.creating.charsequence.contains.CharSequenceContains
 import ch.tutteli.atrium.logic.creating.charsequence.contains.CharSequenceContains.SearchBehaviour
 import ch.tutteli.atrium.logic.creating.charsequence.contains.steps.*
-import kotlin.jvm.JvmMultifileClass
-import kotlin.jvm.JvmName
 
 /**
  * Restricts a `contains` assertion by specifying that the number of occurrences of the value which we are looking
@@ -27,7 +20,7 @@ import kotlin.jvm.JvmName
 infix fun <T : CharSequence, S : SearchBehaviour> CharSequenceContains.EntryPointStep<T, S>.atLeast(
     times: Int
 ): AtLeastCheckerStep<T, S> =
-    _logic.atLeastCheckerStep(times, StaticNames.containsNotValuesFun) { "`${StaticNames.atLeast} $it`" }
+    _logic.atLeastCheckerStep(times, StaticNames.notToContainValuesFun) { "`${StaticNames.atLeast} $it`" }
 
 /**
  * Restricts a `contains at least` assertion by specifying that the number of occurrences of the value which we
@@ -48,7 +41,7 @@ infix fun <T : CharSequence, S : SearchBehaviour> AtLeastCheckerStep<T, S>.butAt
     times: Int
 ): ButAtMostCheckerStep<T, S> = _logic.butAtMostCheckerStep(
     times,
-    StaticNames.containsNotValuesFun,
+    StaticNames.notToContainValuesFun,
     { l, u -> "`${StaticNames.atLeast} $l ${StaticNames.butAtMost} $u`" },
     { "`${StaticNames.atLeast} $it`" },
     { "`${StaticNames.butAtMost} $it`" },
@@ -70,7 +63,7 @@ infix fun <T : CharSequence, S : SearchBehaviour> AtLeastCheckerStep<T, S>.butAt
 infix fun <T : CharSequence, S : SearchBehaviour> CharSequenceContains.EntryPointStep<T, S>.exactly(
     times: Int
 ): ExactlyCheckerStep<T, S> =
-    _logic.exactlyCheckerStep(times, StaticNames.containsNotValuesFun) { "`${StaticNames.exactly} $it`" }
+    _logic.exactlyCheckerStep(times, StaticNames.notToContainValuesFun) { "`${StaticNames.exactly} $it`" }
 
 /**
  * Restricts a `contains` assertion by specifying that the number of occurrences of the value which we
@@ -92,7 +85,7 @@ infix fun <T : CharSequence, S : SearchBehaviour> CharSequenceContains.EntryPoin
     times: Int
 ): AtMostCheckerStep<T, S> = _logic.atMostCheckerStep(
     times,
-    StaticNames.containsNotValuesFun,
+    StaticNames.notToContainValuesFun,
     { "`${StaticNames.atMost} $it`" },
     { "`${StaticNames.atLeast} $it`" },
     { "`${StaticNames.exactly} $it`" }
@@ -112,4 +105,4 @@ infix fun <T : CharSequence, S : SearchBehaviour> CharSequenceContains.EntryPoin
 infix fun <T : CharSequence, S : SearchBehaviour> CharSequenceContains.EntryPointStep<T, S>.notOrAtMost(
     times: Int
 ): NotOrAtMostCheckerStep<T, S> =
-    _logic.notOrAtMostCheckerStep(times, StaticNames.containsNotValuesFun) { "`${StaticNames.notOrAtMost} $it`" }
+    _logic.notOrAtMostCheckerStep(times, StaticNames.notToContainValuesFun) { "`${StaticNames.notOrAtMost} $it`" }
