+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/charSequenceToContainCreators.kt
@@ -1,8 +1,3 @@
-//TODO rename file to charSequenceToContain... in 0.18.0
-//TODO remove JvmMultifileClass with 1.0.0
-@file:JvmMultifileClass
-@file:JvmName("CharSequenceContainsCreatersKt")
-
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.api.infix.en_GB.creating.All
