+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/deprecatedIterableAssertions.kt
@@ -279,62 +279,6 @@ infix fun <E, T : Iterable<E>> Expect<T>.containsNot(expected: E): Expect<T> =
 infix fun <E, T : Iterable<E>> Expect<T>.containsNot(values: Values<E>): Expect<T> =
     it notToContain o the values
 
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Creates an [Expect] for the result of calling `min()` on the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @param o The filler object [o].
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @since 0.12.0
- */
-infix fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.min(@Suppress("UNUSED_PARAMETER") o: o): Expect<E> =
-    _logic.min(::identity).transform()
-
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Expects that the result of calling `min()` on the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.12.0
- */
-infix fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.min(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.min(::identity).collectAndAppend(assertionCreator)
-
-
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Creates an [Expect] for the result of calling `max()` on the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @param o The filler object [o].
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @since 0.12.0
- */
-infix fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.max(@Suppress("UNUSED_PARAMETER") o: o): Expect<E> =
-    _logic.max(::identity).transform()
-
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Expects that the result of calling `max()` on  the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.12.0
- */
-infix fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.max(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.max(::identity).collectAndAppend(assertionCreator)
-
-
 /**
  * Expects that the subject of `this` expectation (an [Iterable]) contains an entry holding
  * the assertions created by [assertionCreatorOrNull] or an entry which is `null` in case [assertionCreatorOrNull]
@@ -425,33 +369,3 @@ infix fun <E, T : Iterable<E>> Expect<T>.hasNot(@Suppress("UNUSED_PARAMETER") ne
 @Suppress("DEPRECATION")
 infix fun <E, T : Iterable<E>> Expect<T>.contains(@Suppress("UNUSED_PARAMETER") noDuplicates: noDuplicates): Expect<T> =
     _logicAppend { containsNoDuplicates(::identity) }
-
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Turns `Expect<E, T : Iterable<E>>` into `Expect<List<E>`.
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature { f(it::asList) }` if you want to show the transformation in reporting.
- *
- * @return The newly created [Expect] for the transformed subject.
- *
- * @since 0.14.0
- */
-infix fun <E, T : Iterable<E>> Expect<T>.asList(
-    @Suppress("UNUSED_PARAMETER") o: o
-): Expect<List<E>> = _logic.changeSubject.unreported { it.toList() }
-
-//TODO 0.18.0 move to iterableExpectations.kt
-/**
- * Expects that the subject of `this` expectation holds all assertions the given [assertionCreator] creates for
- * the subject as [List].
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature of({ f(it::asList) }, assertionCreator)` if you want to show the transformation in reporting.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.14.0
- */
-infix fun <E, T : Iterable<E>> Expect<T>.asList(assertionCreator: Expect<List<E>>.() -> Unit): Expect<T> =
-    apply { asList(o)._logic.appendAsGroup(assertionCreator) }
