+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/deprecatedMapAssertions.kt
@@ -136,17 +136,6 @@ inline infix fun <K, reified V : Any, T : Map<out K, V?>> Expect<T>.contains(key
 inline infix fun <K, reified V : Any, T : Map<out K, V?>> Expect<T>.containsOnly(keyValue: KeyWithValueCreator<K, V>): Expect<T> =
     it toContain o inAny order but only entry keyValue
 
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Helper function to create a [KeyWithValueCreator] based on the given [key] and [valueAssertionCreatorOrNull]
- * -- allows to express `Pair<K, V>, vararg Pair<K, V>`.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.toContainKeyValue
- *
- */
-fun <K, V : Any> keyValue(key: K, valueAssertionCreatorOrNull: (Expect<V>.() -> Unit)?): KeyWithValueCreator<K, V> =
-    KeyWithValueCreator(key, valueAssertionCreatorOrNull)
-
 /**
  * Expects that the subject of `this` expectation (a [Map]) contains for each [KeyWithValueCreator] in [allKeyValues],
  * a key as defined by [KeyWithValueCreator.key] with a corresponding value which either holds all
@@ -254,161 +243,3 @@ infix fun <K, T : Map<out K, *>> Expect<T>.containsKey(key: K): Expect<T> =
 )
 infix fun <K, T : Map<out K, *>> Expect<T>.containsNotKey(key: K): Expect<T> =
     _logicAppend { containsNotKey(::identity, key) }
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the subject of `this` expectation (a [Map]) contains the given [key],
- * creates an [Expect] for the corresponding value and returns the newly created [Expect],
- * so that further fluent calls are expectations about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.getExistingFeature
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.getExisting(key: K): Expect<V> =
-    _logic.getExisting(::identity, key).transform()
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the subject of `this` expectation (a [Map]) contains the given [key] and that
- * the corresponding value holds all assertions the given [KeyWithCreator.assertionCreator] creates for it.
- *
- * @param key Use the function `key(...) { ... }` to create a [KeyWithCreator] where the first parameter corresponds
- *  to the key and the second is the `assertionCreator`-lambda
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.getExisting
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.getExisting(key: KeyWithCreator<K, V>): Expect<T> =
-    _logic.getExisting(::identity, key.key).collectAndAppend(key.assertionCreator)
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Helper function to create an [KeyWithCreator] based on the given [key] and [assertionCreator].
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.getExisting
- *
- */
-fun <K, V> key(key: K, assertionCreator: Expect<V>.() -> Unit): KeyWithCreator<K, V> =
-    KeyWithCreator(key, assertionCreator)
-
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Map.keys] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.keysFeature
- *
- */
-val <K, T : Map<out K, *>> Expect<T>.keys: Expect<Set<K>>
-    get() = _logic.property(Map<out K, *>::keys).transform()
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the property [Map.keys] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.keys
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.keys(assertionCreator: Expect<Set<K>>.() -> Unit): Expect<T> =
-    _logic.property(Map<out K, *>::keys).collectAndAppend(assertionCreator)
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Map.values] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.valuesFeature
- *
- */
-val <V, T : Map<*, V>> Expect<T>.values: Expect<Collection<V>>
-    get() = _logic.property(Map<out Any?, V>::values).transform()
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the property [Map.keys] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.values
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.values(assertionCreator: Expect<Collection<V>>.() -> Unit): Expect<T> =
-    _logic.property(Map<out K, V>::values).collectAndAppend(assertionCreator)
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Turns `Expect<Map<K, V>>` into `Expect<Set<Map.Entry<K, V>>>`.
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature { f(it::entries) }` if you want to show the transformation in reporting.
- *
- * @return The newly created [Expect] for the transformed subject.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.asEntriesFeature
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.asEntries(
-    @Suppress("UNUSED_PARAMETER") o: o
-): Expect<Set<Map.Entry<K, V>>> = _logic.changeSubject.unreported { it.entries }
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Turns `Expect<Map<K, V>>` into `Expect<Set<Map.Entry<K, V>>>` and expects that it holds all assertions the given
- * [assertionCreator] creates for it.
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature { f(it::entries) }` if you want to show the transformation in reporting.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.asEntries
- *
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.asEntries(
-    assertionCreator: Expect<Set<Map.Entry<K, V>>>.() -> Unit
-): Expect<T> = apply { asEntries(o)._logic.appendAsGroup(assertionCreator) }
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the subject of `this` expectation (a [Map]) is an empty [Map].
- *
- * @param empty Use the pseudo-keyword `empty`.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.toBeEmpty
- *
- */
-infix fun <T : Map<*, *>> Expect<T>.toBe(@Suppress("UNUSED_PARAMETER") empty: empty): Expect<T> =
-    _logicAppend { isEmpty(::toEntries) }
-
-//TODO move to mapExpectations.kt with 0.18.0
-/**
- * Expects that the subject of `this` expectation (a [Map]) is not an empty [Map].
- *
- * @param empty Use the pseudo-keyword `empty`.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapExpectationSamples.notToBeEmpty
- *
- */
-infix fun <T : Map<*, *>> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") empty: empty): Expect<T> =
-    _logicAppend { isNotEmpty(::toEntries) }
-
-private fun <T : Map<*, *>> toEntries(t: T): Collection<*> = t.entries
