+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/deprecatedMapEntryAssertions.kt
@@ -1,7 +1,10 @@
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.logic.*
+import ch.tutteli.atrium.logic._logicAppend
+import ch.tutteli.atrium.logic.isKeyValue
+import ch.tutteli.atrium.logic.key
+import ch.tutteli.atrium.logic.value
 
 /**
  * Expects that the property [Map.Entry.key] of the subject of `this` expectation
@@ -18,53 +21,3 @@ import ch.tutteli.atrium.logic.*
 @Deprecated("Use toEqualKeyValue; will be removed with 1.0.0 at the latest", ReplaceWith("this.toEqualKeyValue<K, V, T>(keyValuePair)"))
 infix fun <K, V, T : Map.Entry<K, V>> Expect<T>.isKeyValue(keyValuePair: Pair<K, V>): Expect<T> =
     _logicAppend { isKeyValue(keyValuePair.first, keyValuePair.second) }
-
-//TODO move to mapEntryExpectations with 0.18.0
-/**
- * Creates an [Expect] for the property [Map.Entry.key] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapEntryExpectationSamples.keyFeature
- */
-val <K, T : Map.Entry<K, *>> Expect<T>.key: Expect<K>
-    get() = _logic.key().transform()
-
-//TODO move to mapEntryExpectations with 0.18.0
-/**
- * Expects that the property [Map.Entry.key] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapEntryExpectationSamples.key
- */
-infix fun <K, V, T : Map.Entry<K, V>> Expect<T>.key(assertionCreator: Expect<K>.() -> Unit): Expect<T> =
-    _logic.key().collectAndAppend(assertionCreator)
-
-//TODO move to mapEntryExpectations with 0.18.0
-/**
- * Creates an [Expect] for the property [Map.Entry.value] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapEntryExpectationSamples.valueFeature
- */
-val <V, T : Map.Entry<*, V>> Expect<T>.value: Expect<V>
-    get() = _logic.value().transform()
-
-//TODO move to mapEntryExpectations with 0.18.0
-/**
- * Expects that the property [Map.Entry.value] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.MapEntryExpectationSamples.value
- */
-infix fun <K, V, T : Map.Entry<K, V>> Expect<T>.value(assertionCreator: Expect<V>.() -> Unit): Expect<T> =
-    _logic.value().collectAndAppend(assertionCreator)
