+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/iterableSubjectChangers.kt
@@ -1,0 +1,34 @@
+package ch.tutteli.atrium.api.infix.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic.*
+import kotlin.jvm.JvmMultifileClass
+import kotlin.jvm.JvmName
+
+/**
+ * Turns `Expect<E, T : Iterable<E>>` into `Expect<List<E>`.
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature { f(it::asList) }` if you want to show the transformation in reporting.
+ *
+ * @return The newly created [Expect] for the transformed subject.
+ *
+ * @since 0.14.0
+ */
+infix fun <E, T : Iterable<E>> Expect<T>.asList(
+    @Suppress("UNUSED_PARAMETER") o: o
+): Expect<List<E>> = _logic.changeSubject.unreported { it.toList() }
+
+/**
+ * Expects that the subject of `this` expectation holds all assertions the given [assertionCreator] creates for
+ * the subject as [List].
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature of({ f(it::asList) }, assertionCreator)` if you want to show the transformation in reporting.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @since 0.14.0
+ */
+infix fun <E, T : Iterable<E>> Expect<T>.asList(assertionCreator: Expect<List<E>>.() -> Unit): Expect<T> =
+    apply { asList(o)._logic.appendAsGroup(assertionCreator) }
