+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/listFeatureExtractors.kt
@@ -1,4 +1,3 @@
-//TODO rename file to listExpectations.kt in 0.18.0
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.api.infix.en_GB.creating.IndexWithCreator
