+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/sequenceSubjectChangers.kt
@@ -1,4 +1,3 @@
-//TODO rename file to sequenceExpectations in 0.18.0
 package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.creating.Expect
