+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/throwableFeatureExtractors.kt
@@ -1,7 +1,6 @@
-package ch.tutteli.atrium.api.fluent.en_GB
+package ch.tutteli.atrium.api.infix.en_GB
 
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.logic.creating.typeutils.CharSequenceOrNumberOrChar
 import ch.tutteli.atrium.logic._logic
 import ch.tutteli.atrium.logic.causeIsA
 import ch.tutteli.atrium.logic.creating.transformers.SubjectChangerBuilder
@@ -13,10 +12,10 @@ import kotlin.reflect.KClass
  *
  * @return The newly created [Expect] for the property [Throwable.message] of the subject of `this` expectation.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableFeatureExtractorSamples.message
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.ThrowableExpectationSamples.messageFeature
  */
 val <T : Throwable> Expect<T>.message: Expect<String>
-    get() = feature(Throwable::message).notToEqualNull()
+    get() = it feature Throwable::message notToEqualNull o
 
 /**
  * Expects that the property [Throwable.message] of the subject of `this` expectation is not null and
@@ -25,40 +24,20 @@ val <T : Throwable> Expect<T>.message: Expect<String>
  *
  * @return an [Expect] for the subject of `this` expectation.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableFeatureExtractorSamples.messageFeature
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.ThrowableExpectationSamples.message
  */
-fun <T : Throwable> Expect<T>.message(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
-    feature(Throwable::message) { notToEqualNull(assertionCreator) }
-
-/**
- * Expects that the property [Throwable.message] of the subject of `this` expectation is not null and contains
- * [expected]'s [toString] representation and the [toString] representation of the [otherExpected] (if given),
- * using a non-disjoint search.
- *
- * Notice that a runtime check applies which assures that only [CharSequence], [Number] and [Char] are passed.
- * This function expects [CharSequenceOrNumberOrChar] (which is a typealias for [Any]) for your convenience,
- * so that you can mix [String] and [Int] for instance.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableFeatureExtractorSamples.messageToContain
- *
- * @since 0.17.0
- */
-fun <T : Throwable> Expect<T>.messageToContain(
-    expected: CharSequenceOrNumberOrChar,
-    vararg otherExpected: CharSequenceOrNumberOrChar
-): Expect<T> = message { toContain(expected, *otherExpected) }
+infix fun <T : Throwable> Expect<T>.message(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
+    it feature of(Throwable::message) { it notToEqualNull assertionCreator }
 
 /**
  * Expects that the property [Throwable.cause] of the subject *is a* [TExpected] (the same type or a sub-type),
  * creates an [Expect] of the [TExpected] type for it and returns it.
  *
- * @return The newly created [Expect] for the property [Throwable.cause] of the subject of `this` expectation.
+ * @return The newly created [Expect] for the property [Throwable.cause] of the subject of `this` expectation
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableFeatureExtractorSamples.causeFeature
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.ThrowableExpectationSamples.causeFeature
  *
- * @since 0.10.0
+ * @since 0.12.0
  */
 inline fun <reified TExpected : Throwable> Expect<out Throwable>.cause(): Expect<TExpected> =
     causeToBeAnInstanceOf(TExpected::class).transform()
@@ -68,20 +47,21 @@ internal fun <TExpected : Throwable> Expect<out Throwable>.causeToBeAnInstanceOf
     kClass: KClass<TExpected>
 ): SubjectChangerBuilder.ExecutionStep<Throwable?, TExpected> = _logic.causeIsA(kClass)
 
+
 /**
  *
  * Expects that the property [Throwable.cause] of the subject *is a* [TExpected] (the same type or a sub-type) and
  * holds all assertions the given [assertionCreator] creates for it.
  *
- * Notice, in contrast to other expectation functions which expect an [assertionCreator], this function returns not
+ * Notice, in contrast to other assertion functions which expect an [assertionCreator], this function returns not
  * [Expect] of the initial type, which was some type `T `, but an [Expect] of the specified type [TExpected].
  *
  * @return an [Expect] for the subject of `this` expectation.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ThrowableFeatureExtractorSamples.cause
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.ThrowableExpectationSamples.cause
  *
- * @since 0.10.0
+ * @since 0.12.0
  */
-inline fun <reified TExpected : Throwable> Expect<out Throwable>.cause(
+inline infix fun <reified TExpected : Throwable> Expect<out Throwable>.cause(
     noinline assertionCreator: Expect<TExpected>.() -> Unit
 ): Expect<TExpected> = causeToBeAnInstanceOf(TExpected::class).transformAndAppend(assertionCreator)
